backend/requirements.txt
aiofiles==23.2.1
anyio==4.3.0
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
exceptiongroup==1.2.0
fastapi==0.110.0
h11==0.14.0
idna==3.6
pycparser==2.21
pydantic==2.6.4
PyJWT==2.8.0
python-multipart==0.0.9
sniffio==1.3.1
starlette==0.36.3
typing-extensions==4.10.0
uvicorn==0.28.0
